refactor(background): refactors the electron background process
splits out the giant file into smaller files! hopefully it's easier to maintain now...
