+++ src/background/background-constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  APP_SCHEME: "app"
+};
