+++ src/background/open-file.js
@@ -1,0 +1,8 @@
+import { app } from "electron";
+import { windows } from "./windows";
+
+export function openFile(filePath) {
+  app.addRecentDocument(filePath);
+
+  windows["mainWindow"].webContents.send("open-preset", filePath);
+}
