+++ vue.config.js
@@ -39,6 +39,7 @@ module.exports = {
 
   pluginOptions: {
     electronBuilder: {
+      mainProcessFile: "src/background/background.js",
       nodeIntegration: true,
 
       externals: [
